Modify src files to optimize genome sequencing and create .paf with reads and references

+++ src/alignment.cpp
@@ -14,17 +14,17 @@ int Align(
     std::string* cigar,
     unsigned int* target_begin)
 {
-    unsigned int n = query_len;
-    unsigned int m = target_len;
+    size_t n = query_len;
+    size_t m = target_len;
 
     // Vettore 1D per simulare la matrice 2D (Cache-friendly)
     std::vector<int> H((n + 1) * (m + 1), 0);
     std::vector<char> trace((n + 1) * (m + 1), 0); // 'D'=Diag, 'U'=Up, 'L'=Left
 
-    auto idx = [&](unsigned int i, unsigned int j) { return i * (m + 1) + j; };
+    auto idx = [&](size_t i, size_t j) { return i * (m + 1) + j; };
 
     // Inizializzazione della matrice in base al tipo di allineamento
-    for (unsigned int i = 1; i <= n; ++i) {
+    for (size_t i = 1; i <= n; ++i) {
         if (type == AlignmentType::Global) {
             H[idx(i, 0)] = i * gap;
             trace[idx(i, 0)] = 'U';
@@ -32,7 +32,7 @@ int Align(
             H[idx(i, 0)] = 0;
         }
     }
-    for (unsigned int j = 1; j <= m; ++j) {
+    for (size_t j = 1; j <= m; ++j) {
         if (type == AlignmentType::Global) {
             H[idx(0, j)] = j * gap;
             trace[idx(0, j)] = 'L';
@@ -42,11 +42,11 @@ int Align(
     }
 
     int max_score = -999999;
-    unsigned int max_i = 0, max_j = 0;
+    size_t max_i = 0, max_j = 0;
 
     // Popolamento della matrice
-    for (unsigned int i = 1; i <= n; ++i) {
-        for (unsigned int j = 1; j <= m; ++j) {
+    for (size_t i = 1; i <= n; ++i) {
+        for (size_t j = 1; j <= m; ++j) {
             int score_diag = H[idx(i - 1, j - 1)] + (query[i - 1] == target[j - 1] ? match : mismatch);
             int score_up   = H[idx(i - 1, j)] + gap;
             int score_left = H[idx(i, j - 1)] + gap;
@@ -82,7 +82,7 @@ int Align(
     }
 
     // Ricerca del punto di inizio per il Backtracking
-    unsigned int start_i = n, start_j = m;
+    size_t start_i = n, start_j = m;
     int final_score = 0;
 
     if (type == AlignmentType::Global) {
@@ -94,13 +94,13 @@ int Align(
     } else if (type == AlignmentType::SemiGlobal) {
         final_score = H[idx(n, 0)];
         start_i = n; start_j = 0;
-        for (unsigned int j = 1; j <= m; ++j) {
+        for (size_t j = 1; j <= m; ++j) {
             if (H[idx(n, j)] > final_score) {
                 final_score = H[idx(n, j)];
                 start_i = n; start_j = j;
             }
         }
-        for (unsigned int i = 1; i <= n; ++i) {
+        for (size_t i = 1; i <= n; ++i) {
             if (H[idx(i, m)] > final_score) {
                 final_score = H[idx(i, m)];
                 start_i = i; start_j = m;
@@ -109,8 +109,8 @@ int Align(
     }
 
     // Backtracking per calcolare la CIGAR e la posizione di inizio
-    unsigned int curr_i = start_i;
-    unsigned int curr_j = start_j;
+    size_t curr_i = start_i;
+    size_t curr_j = start_j;
     std::string raw_cigar = "";
 
     while (curr_i > 0 || curr_j > 0) {
